requirements.txt
flask
openai
python-dotenv
numpy
python-frontmatter
pyyaml
pymorphy3